@@ -13,48 +13,48 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^6.1.3",
-    "@angular/common": "^6.1.3",
-    "@angular/compiler": "^6.1.3",
-    "@angular/core": "^6.1.3",
-    "@angular/forms": "^6.1.3",
-    "@angular/http": "^6.1.3",
-    "@angular/material": "^6.4.5",
-    "@angular/platform-browser": "^6.1.3",
-    "@angular/platform-browser-dynamic": "^6.1.3",
-    "@angular/platform-server": "^6.1.3",
-    "@angular/router": "^6.1.3",
-    "@nguniversal/module-map-ngfactory-loader": "^6.0.0",
+    "@angular/animations": "^7.0.2",
+    "@angular/common": "^7.0.2",
+    "@angular/compiler": "^7.0.2",
+    "@angular/core": "^7.0.2",
+    "@angular/forms": "^7.0.2",
+    "@angular/http": "^7.0.2",
+    "@angular/material": "^7.0.2",
+    "@angular/platform-browser": "^7.0.2",
+    "@angular/platform-browser-dynamic": "^7.0.2",
+    "@angular/platform-server": "^7.0.2",
+    "@angular/router": "^7.0.2",
+    "@nguniversal/module-map-ngfactory-loader": "^7.0.2",
     "aspnet-prerendering": "^3.0.1",
     "bootstrap": "^3.3.7",
     "core-js": "^2.5.7",
-    "rxjs": "^6.2.2",
-    "rxjs-compat": "^6.2.2",
+    "rxjs": "^6.3.3",
+    "rxjs-compat": "^6.3.3",
     "zone.js": "^0.8.26",
-    "@angular/cdk": "^6.2.0"
+    "@angular/cdk": "^7.0.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.7.0",
-    "@angular/cli": "~6.1.4",
-    "@angular/compiler-cli": "^6.1.3",
-    "@angular/language-service": "^6.1.3",
-    "@types/jasmine": "~2.8.8",
-    "@types/jasminewd2": "~2.0.3",
-    "@types/node": "~10.7.1",
-    "codelyzer": "^4.4.4",
-    "jasmine-core": "~3.2.1",
+    "@angular-devkit/build-angular": "~0.10.4",
+    "@angular/cli": "~7.0.4",
+    "@angular/compiler-cli": "^7.0.2",
+    "@angular/language-service": "^7.0.2",
+    "@types/jasmine": "~2.8.9",
+    "@types/jasminewd2": "~2.0.5",
+    "@types/node": "~10.12.2",
+    "codelyzer": "^4.5.0",
+    "jasmine-core": "~3.3.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~3.0.0",
+    "karma": "~3.1.1",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "^2.0.1",
+    "karma-coverage-istanbul-reporter": "^2.0.4",
     "karma-jasmine": "~1.1.2",
-    "karma-jasmine-html-reporter": "^1.3.0",
-    "protractor": "~5.4.0",
+    "karma-jasmine-html-reporter": "^1.4.0",
+    "protractor": "~5.4.1",
     "ts-node": "~7.0.1",
     "tslint": "~5.11.0",
-    "typescript": "~2.9.2"
+    "typescript": "~3.1.6"
   },
   "optionalDependencies": {
-    "node-sass": "^4.9.3"
+    "node-sass": "^4.10.0"
   }
 }
