@@ -16,14 +16,16 @@ namespace translate_spa.Actions
     {
         private readonly Translation _translation;
         private readonly TranslateService _translateService;
+        private readonly static string _googleApiKey = Startup.Configuration.GetSection("GoogleTranslateApi")["ApiKey"];
+        private readonly static string _googleApplicationName = Startup.Configuration.GetSection("GoogleTranslateApi")["ApplicationName"];
 
         public GoogleTranslate(Translation translation)
         {
             _translation = translation;
             _translateService = new TranslateService(new BaseClientService.Initializer
             {
-                ApplicationName = "esignatur translations portal",
-                    ApiKey = "AIzaSyDAWis6FJJerxydhqV-iPSjChj3cY4E1FQ",
+                ApplicationName = _googleApplicationName,
+                    ApiKey = _googleApiKey,
             });
         }
 
