@@ -1,7 +1,6 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Linq.Expressions;
 using System.Threading.Tasks;
 
 using Microsoft.AspNetCore.Http;
@@ -13,11 +12,12 @@ using Serilog;
 using translate_spa.Actions;
 using translate_spa.Controllers.ActionFilters;
 using translate_spa.Models;
-using translate_spa.Models.Interfaces;
 using translate_spa.MongoDB;
 using translate_spa.MongoDB.DbBuilder;
 using translate_spa.Querys;
 using translate_spa.Repositories;
+using translate_spa.Tasks;
+// using translate_spa.Tasks;
 using translate_spa.Utilities;
 
 namespace translate_spa.Controllers
@@ -48,7 +48,8 @@ namespace translate_spa.Controllers
 
             var mongoRepository = new MongoRepository<Translation>(new BaseDbBuilder());
 
-            return new AddTranslation(mongoRepository).Execute(translation)as Translation;
+            new SetEnvironmentFromKey(translation).Execute();
+            return new AddTranslation(mongoRepository).Execute(translation) as Translation;
         }
 
         [HttpPost("~/api/[controller]/[action]")]
@@ -80,6 +81,8 @@ namespace translate_spa.Controllers
             {
                 throw new Exception($"Adding dublicate translation-key not allowed. the used translationkey are allready in use");
             }
+
+            new SetEnvironmentFromKey(translation).Execute();
 
             new UpdateTranslation(mongoRepository).Execute(translation);
 
@@ -130,6 +133,42 @@ namespace translate_spa.Controllers
             return queryResult;
         }
 
+        [HttpGet("~/api/[controller]/[action]/{env}")]
+        public async Task<IEnumerable<OldTranslation>> OldQuery(TranslationsEnvironment env)
+        {
+            var mongoRepository = new MongoRepository<Translation>(new BaseDbBuilder());
+            var queryExpression = PredicateBuilder.True<Translation>();
+
+            queryExpression = queryExpression.And(x => x.Environment == env);
+
+            if (_Branch.HasBranch)
+            {
+                Log.Debug($"Query: adding Branch predicate for branch: '{_Branch.Value}'");
+
+                queryExpression = queryExpression.And(x => x.Branch == null);
+            }
+
+            if (env != TranslationsEnvironment.Desktop && env != TranslationsEnvironment.OldDesktop)
+            {
+                queryExpression = queryExpression.Or(x => x.Environment == TranslationsEnvironment.Common);
+            }
+
+            var queryResult = await mongoRepository.QueryAsync(queryExpression);
+
+            Log.Debug($"Query: returning '{queryResult.Count()}' environment '{env.ToString()}''");
+
+            var result = queryResult.Select(x => new OldTranslation()
+            {
+                KEY = x.Key,
+                DA = x.Da,
+                EN = x.En,
+                SV = x.Sv,
+                NB = x.Nb,
+            });
+
+            return result;
+        }
+
         [HttpGet("~/api/[controller]/[action]/{env}/{lang}")]
         public async Task<JsonResult> AngularQuery(Language lang, string env)
         {
@@ -145,8 +184,16 @@ namespace translate_spa.Controllers
             predicate = predicate.And(x => x.Environment == environment.Value);
 
             var result = mongoRepository.Query(predicate);
+
+            // Log.Debug($"AngularQuery: returning '{result.Count()}' entries for language'{lang}' on environment '{environment.Value}'");
+            // var resultDictionary = result.ToDictionary(t => t.Key, t => t.GetByLanguage(lang));
+
             Log.Debug($"AngularQuery: returning '{result.Count()}' entries for language'{lang}' on environment '{environment.Value}'");
-            var resultDictionary = result.ToDictionary(t => t.Key, t => t.GetByLanguage(lang));
+
+            // TODO: maybe this should be changed, at least for the desktop, we skip "Desktop.", but maybe we should skip "Desktop.App." on the 
+            // new desktop, and i dont know, what we should do on the web.
+            // HENCE, maybe create a function for this, and may enable skip on 'dot' via appsettings.
+            var resultDictionary = result.ToDictionary(t => string.Join(".", t.Key.Split(".").Skip(1)), t => t.GetByLanguage(lang));
 
             var jsonResult = JsonHelper.Unflatten(resultDictionary);
 
@@ -182,11 +229,11 @@ namespace translate_spa.Controllers
             var translated = translations.Select(x => new Translation()
             {
                 Key = x.KEY,
-                    Da = x.DA,
-                    En = x.EN,
-                    Sv = x.SV,
-                    Nb = x.NB,
-                    Branch = string.Empty
+                Da = x.DA,
+                En = x.EN,
+                Sv = x.SV,
+                Nb = x.NB,
+                Branch = string.Empty
             });
 
             var mongoRepository = new MongoRepository<Translation>(new BaseDbBuilder());
@@ -247,8 +294,10 @@ namespace translate_spa.Controllers
         }
 
         [HttpGet("~/api/[controller]/[action]")]
-        public async Task<IEnumerable<Translation>> GoogleTranslate()
+        public async Task Notify()
         {
+            /*    
+            // Translate all:
             var mongoRepository = new MongoRepository<Translation>(new BaseDbBuilder());
             var result = mongoRepository.All()
                 .Where(x => string.IsNullOrEmpty(x.Branch) &&
@@ -257,12 +306,16 @@ namespace translate_spa.Controllers
                         string.IsNullOrEmpty(x.Nb)) &&
                     x.Da.Split(' ', StringSplitOptions.RemoveEmptyEntries).Count() <= 3);
 
-            /* foreach (var item in result)
-            {
-                new GoogleTranslate(item, Log).Execute();
-            } */
-
             return result.Select(x => new GoogleTranslate(x).Execute());
+            */
+            var mongoRepository = new MongoRepository<Translation>(new BaseDbBuilder());
+
+            var translations = mongoRepository.All()
+                .Where(x => string.IsNullOrEmpty(x.Branch) &&
+                    x.HasMissingTranslation()).ToList();
+
+            Log.Debug($"Running missing translations task. Missing translations: {translations.Count()}");
+            await new MissingTranslationsTask(translations).ExecuteAsync();
         }
     }
 }
